Refactor DownloadConfig
