+++ src/datasets/commands/run_beam.py
@@ -7,7 +7,8 @@ from typing import List
 from datasets import config
 from datasets.builder import DatasetBuilder
 from datasets.commands import BaseDatasetsCLICommand
-from datasets.download.download_manager import DownloadConfig, DownloadMode
+from datasets.download.download_config import DownloadConfig
+from datasets.download.download_manager import DownloadMode
 from datasets.load import dataset_module_factory, import_main_class
