+++ src/datasets/download/__init__.py
@@ -1,9 +1,10 @@
 __all__ = [
-    # "DownloadConfig",
+    "DownloadConfig",
     "DownloadManager",
     "DownloadMode",
     "StreamingDownloadManager",
 ]
 
+from .download_config import DownloadConfig
 from .download_manager import DownloadManager, DownloadMode
 from .streaming_download_manager import StreamingDownloadManager
