+++ tests/test_download_manager.py
@@ -4,8 +4,9 @@ from pathlib import Path
 
 import pytest
 
+from datasets.download.download_config import DownloadConfig
 from datasets.download.download_manager import DownloadManager
-from datasets.utils.file_utils import DownloadConfig, hash_url_to_filename
+from datasets.utils.file_utils import hash_url_to_filename
 
 
 URL = "http://www.mocksite.com/file1.txt"
