+++ tests/test_dummy_data_autogenerate.py
@@ -6,10 +6,10 @@ from unittest import TestCase
 import datasets.config
 from datasets.builder import GeneratorBasedBuilder
 from datasets.commands.dummy_data import DummyDataGeneratorDownloadManager, MockDownloadManager
+from datasets.download.download_config import DownloadConfig
 from datasets.features import Features, Value
 from datasets.info import DatasetInfo
 from datasets.splits import Split, SplitGenerator
-from datasets.utils.file_utils import DownloadConfig
 from datasets.utils.version import Version
